+++ 06labs/clock_output/clock.h
@@ -1,7 +1,6 @@
 #ifndef CLOCK_H
 #define CLOCK_H
-#include <stdint.h>
 
-void op_sysclk(uint8_t div);
+void op_sysclk(unsigned int div);
 
 #endif
