+++ 06labs/clock_pll/clock.h
@@ -1,8 +1,7 @@
 #ifndef CLOCK_H
 #define CLOCK_H
-#include <stdint.h>
 
-void op_sysclk(uint8_t div);
+void op_sysclk(unsigned int div);
 void set_sysclk_pll(void);
 
 #endif
